third_party/blink/web_tests/http/tests/security/sandbox-iframe-allows-modals-expected.txt

ALERT: Yay!
CONFIRM: Question?
PROMPT: Question?, default text: 
This is a testharness.js-based test.
Harness: the test ran to completion.
